Make it compile with VS2013

+++ src/alone/binarize.c
@@ -26,7 +26,7 @@
 
 static void VS_CC set_lut(alone_t *ah, int thresh, uint16_t val0, uint16_t val1)
 {
-    size_t size = ah->lut_size;
+    int size = (int)ah->lut_size;
     uint16_t *lut = ah->lut;
 
     int i = 0;
