+++ src/alone/limiter.c
@@ -27,7 +27,7 @@
 static void VS_CC set_lut(alone_t *ah, int min, int max)
 {
     uint16_t *lut = ah->lut;
-    size_t size = ah->lut_size;
+    int size = (int)ah->lut_size;
 
     int i = 0;
     while (i < min) lut[i++] = min;
